+++ examples/resources/metabase_glossary_term/resource.tf
@@ -1,0 +1,4 @@
+resource "metabase_glossary_term" "nmv" {
+  term       = "NMV"
+  definition = "Net merchandise value: the total paid by customers, excluding tips."
+}